sound/soc/fsl/fsl_sai: Add clock provider for MCLK output
This allows probing to commence in order where the codec input MCLK
is sourced from the audio cpu hardware. It also prevents sudden removal of
the clock due to PM runtime suspend.
If DT is set up to provide MCLK out, we need to maintain transmitter
running even if not streaming since the enable bit is gating the MCLK
output.